[site] Add dynamic section table
Shortcode for inserting a generated table of content for the current
folder.
This can be used in Markdown files in order to have an always up-to-date
list of the content of the folder.

Supports a "type" parameter which can be set to "section" in order to
list only subsection. Will ignore other content such as normal pages.
Parameter "depth" controls the level to which the content is listed.
Defaults to two if not set.

Fixes lowrisc/opentitan#1419

Signed-off-by: Tobias Wölfel <tobias.woelfel@mailbox.org>