[doc] Use standards-compliant mode for HTML
Before, the pages were using quirks mode. That was visible in a couple
areas, mainly in the font sizes: fonts in tables were overwritten by the
browser to be 12px, while the main text was 16px.

Since we thankfully don't live in IE6 times any more, it's about time
for quirks mode to go.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>