[prim_util] Do not use $clog2() in Xcelium
Xcelium seems to have trouble with the use of $clog2() in
`prim_util_pkg::vbits()`. Fortunately, we can provide a drop-in
replacement which keeps the interface identical.

This approach allows us to keep the interface/function signature
identical, with the option to remove the workaround once we understand
the Xcelium issue better.

If we switched to a macro we had to change all call sites, and make
`prim_util` an include file again.

Fixes #2597

Signed-off-by: Philipp Wagner <phw@lowrisc.org>