[dv/alert_handler] update the ping tests to hit more corner cases
The original `aler_handler_ping_rsp_vseq` will enable the ping timeout
failure and let stress_all and reset tests try to randomly hit all the
corner cases. But some issues for this tests:
1). Long run time. Usually around 2 hours.
2). Could hardly hit any corner cases:
    a). ping interrupted by a reset
    b). ping interrupted by real escalation signal

To try to cover all the corner cases with a reasonable amount of time,
this PR:
1). Modify the ping_rsp_fail vseq to intentionally hit the corner cases,
and to avoid conditions where ping_failure and check_interrupts happened
in the same cycle (this makes scb very hard to predict)
2). Modify LFSR input `wait_cyc_mask_i` to force design to send out ping
request more frequently

The following PR will add a functional coverage for all alert and
escalator to ensure they receiver a ping request. And this coverage will
be disabled when running the `alert_handler_ping_corner_cases` test.

Signed-off-by: Cindy Chen <chencindy@google.com>