[sw] Use Verilator Stop Address
When run under Verilator, the core has an address to write to which can
stop execution. This has not been used by software until now, for no
good reason.

This commit adds a device-specific definition for this address, and
instantiates it with the correct address on verilator. We use the value
of zero to denote that this address doesn't exist on other device
configurations.

The core may write one of the following values:
- `0xFFFFFFFE`, if we reach the end of the Boot ROM (i.e. if the flash
  entry function ever returns, which it shouldn't).
- `0xFFFFFFFF`, if `abort` is called.
- The return value of `main` in the Flash Image. This is likely to be
  either `0x0` (for success), or `0x1` (for failure) to match Unix exit
  codes.

The simulator does not care which value is written, just that a write
hits a specific address.

Closes #501.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>