[otbn] Expand the three writeback variants of bn.mulqacc in docs
Before this change, an operand was optional or required, depending on
the value of another operand. Now, the writeback variant on which this
depended is encoded in the mnemonic.

We could go further, duplicating these into 6 mnemonics to get rid of
the <zero_acc> argument, but I feel like that adds bloat without any
massive advantage.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>