[otbn] Properly parse optional operands for an instruction
This shouldn't have any impact on generated documentation, but the
YAML parser now actually understands what syntax like this means:

    <op0>[, <op1>][, X<op2>]

It should match any of:

    r0
    r0, r1
    r0, Xr2
    r0, r1, Xr2

We don't support nesting brackets (so there are always 2^k possible
syntaxes for some k). We also don't check for ambiguity, so the
ambiguous:

    <op0>[, <op1>][, <op2>]

is accepted. In practice, we're going to parse assembly syntax with a
regex, at least at first, which means the argument will be assigned to
op1. But probably don't do that: it's rather confusing!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>